core/module: simplify code with just-in-time module registering
This patch rework the macro 'MODULE', that used to generate the name of
the global variable containing the module pointer (that could be null if
the module was not registered yet), so it calls a function
that *always* returns the module pointer.

Before:
~~~~~~~
 /* Module pointer. */
 module_t *foo_module;

 /* Pointer on module pointer: used to palliate the fact 'foo_module'
  * may be still null when we try to register a dependency on it. */
 module_t **foo_module_ptr = &foo_module;

 /* MODULE(foo_module) == foo_module */

Now:
~~~~
 /* Always return a pointer, register the module if needed. The module
  * registration is completed with function 'module_implement()'. */
 module_t *foo_module(void);

 /* MODULE(foo_module) == foo_module() */

This change allows us to remove a considerable amount of code that was
there only to manage the fact that we could register dependencies for
modules that didn't exist yet.

What changed:
~~~~~~~~~~~~~
The function 'module_register' still exists but changed purpose: it simply
creates the module and registers it in the internal module QM. A new
function 'module_implement' is created: it allows to register the
constructor, destructor so as a built-in dependency (not several anymore
as it wasn't needed) for the module.

Consequence in code:
~~~~~~~~~~~~~~~~~~~~
"Local modules", for which the implementation (MODULE_BEGIN ...
MODULE_END) isn't preceded by a declaration (MODULE_DECLARE) will have to
be declared as static.

Except for this first point, the patch shouldn't make difference for
normal uses of the modules as the high-level macros are not changed, but
the advanced cases in which the low-level functions are called directly
will probably have to be reworked.

Note about changes in zchk-module.c:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Some modules were registered twice. That isn't tolerated anymore because
a new assertion ensures that it does not happen. Consequently, most uses
of 'Z_MODULE_REGISTER' had to be removed from file 'zchk-module.c'.

Refs: #67486
Change-Id: I18f68795284fcb2ed8384897c230aaf000dc7913