waf: Avoid creating multiple check tasks on the same file.
If a .c file is present in multiple task generators (which is bad because
the same file is compiled multiple times, but sometimes cannot be
avoided), then we were creating multiple tasks with the exact same
parameters.

This triggered warnings in waf, and caused the checks to be done at each
compilation even when they were already done in a previous build.

Avoid this situation thanks to a global set.

Change-Id: Iad140ab1ba4cccbe59c69be8fb1664e7106fc40b