qps: Have per QPS instance loggers.
All the logs emitted by the QPS library did not have the context of which
QPS was concerned. It was really annoying in the QKV database for example,
because it uses one QPS per shard, so we did not have an easy way to know
in which shard a problem occurred.

So enhance the QPS library by having dedicated loggers per QPS instance,
and use them every time it's possible.

The name that is be used for the logger of a QPS instance is given as
argument to qps_open & co. Of course, it must be unique.

Change-Id: Ic5df49225fbce93ab4e12078fad2cc26bab14913