log: Duplicate name in logger_init.
The name given to logger_init can be given on the t_stack for example, so
not duplicating it is not safe.

By chance (or maybe was it intended?), __logger_refresh recomputes and
duplicates the logger->name... in *almost* all codepaths. I'm not sure
about the other codepaths, so don't play with fire: duplicate it.

Note that logger_compute_fullname had to be fixed in order to not leak the
logger name in a codepath. But to me, it was buggy before.

Change-Id: Ifad6608d606b4d0d449478df730da4a03a9f846a