ssl: fix expected length in ssl encrypt / decrypt
According to the doc:
* EVP_EncryptUpdate needs inl + cipher_block_size - 1 in the out buffer
* EVP_EncryptFinal needs cipher_block_size
* EVP_DecryptUpdate needs inl + cipher_block_size

where inl (in length) is data.len for us.

In practice, the worst case is there data.len is a multiple of the block
size, in that case, the padding adds an extra block, so data.len + block
size is sufficient for all cases at encryption.
For decryption, data.len - 1 should be enough but let's do as the doc
says.

Besides to do sb_growlen(out, clen), it's sufficient to have
(sb_avail(out) == clen). The trailing 0 is not counted there.

Finally fix missing sb_grow for the special case where data.len == 0. And
use sb_t instead of SB_1k in the Z_TEST to properly test that case.

Change-Id: I583c5fee892f0f20384941e59738b4f276ada018
Refs: #66644