zchk: do not use Z_GROUP mutable variables in Z_TESTs.
In some tests, we used to define mutable variables at the top of the
Z_GROUP and use them in the Z_TESTs.

This is not a good thing because that means that the Z_TESTs are
potentially not independent.
Plus, this will not work for the upcoming Z_TEST fix for clang.

The solution is to move the mutable variables from the Z_GROUP to the
Z_TESTs that are actually using them.

Change-Id: I6a16725ee9ec7a6fa2e82b864b6f6c8c28345aee
Refs: #68530