z: use blocks instead for Z_TEST.
Clang-7 takes a very long time to compile very long functions.

This is happening in our tests since a Z_GROUP is one big function and
the Z_TESTs are just some scopes in the function.

For example with zchk-iop.c, on default profile:
 - clang-7.0.0 took 3min15s to compile the file.
 - gcc-6.3.0 took 28s to compile the file.

The solution is to use blocks for each Z_TEST. This way, each step is a
separate function, and clang is happier.
We do the same for GCC with nested functions that is a GNU extension.

With zchk-iop.c, on default profile:
 - clang-7.0.0 now takes 6.5s to compile the file.
 - gcc-6.3.0 now takes 8.2s to compile the file.

Change-Id: I824365a410b94a4b85c1c98a2bd357d0f2314ab2
Closes: #68530