el: fix leak in case of process exit on el sigchild callbacks.
In el_sigchld_hook(), the callback called by el_child_fire or the module
method at_fork_on_child_terminated may change the global qm `_G.childs`,
so we cannot call them while iterating on the qm and removing elements
from it.

We first need to list the dead child processes, and then call
el_child_fire() or the module method at_fork_on_child_terminated in
another loop.

To do that, we used a local qv that was init at the beginning of
el_sigchld_hook() and wiped and the end of the function.

Unfortunately, it is possible that process exit in the callbacks called
by el_child_fire or in the module method at_fork_on_child_terminated.

In that case, the local qv was leaked.

The solution is to allocate the local qv in a t_scope that is sealed
before calling the callbacks.
The t_stack is automatically wiped on process exit, so we don't leak
the qv.

Change-Id: Id2fc955806408fc1aaa63b375aaeeca663f5573b