iop-yaml: add IOP packing/unpacking
This is a big commit that introduces packing and unpacking of IOP
objects into and from YAML.

The code is deeply, deeply duplicated from iop-json.blk. A lot of
factorization will come in the following commits. The goal was to get a
minimal working prototype in this commit, so that the following commits
will be much smaller, thus much more reviewable and easy to cover with
unit-tests.

Change-Id: Ib0ad017636f4cf191649b37a69f1a002c9ac3cfd