waf: add SHARED_LIBRARY_SANITIZER to compile shared libraries with sanitizers.
The use of sanitizers was deactivated because it is not possible to
simply compile a shared library with a sanitizer and use it with a
process that is not itself compiled with the sanitizer.
It will complain about missing symbols.

So, it is not possible to simply compile a shared library and use it
with a standard process, for example iopy with python.

However, it is possible to set `LD_PRELOAD` with the sanitizer shared
library to load it when running the process.

Example:
lib-common$ LD_PRELOAD=libclang_rt.asan-x86_64.so python iopy/z_iopy.py

Or, it is possible to directly compile the process with the sanitizer.

To be able to use these possibilities, we need to add a new environment
variable `SHARED_LIBRARY_SANITIZER`, that when set, waf will also
compile the shared libraries with the sanitizer.
This option is of course only available for profiles that use
sanitizers, a.k.a asan and tsan.

Change-Id: Ie6533a6572859ca7f2273784d78f7512739c6d0b