core-mem-stack: fix heap-use-after-free on all_pools at fork.
When the process is forked, the t_pool_g thread local variable is
recreated.
Unfortunately, we didn't reset _G.all_pools, so we kept an invalid
reference to the old t_pool_g variable and when the new one is created,
we have a heap-use-after-free.

The solution is to reset to _G.all_pools at child fork so we don't keep
invalid references to old pools.

Change-Id: I3e2e1e9712240a3535aa33b1e88dc55aad1a443c