qps: add parameter to open a QPS without mapping pages
(cherry picked from commit a4e284598d0b6921ab086dc56602e7bc389832dd)

A new parameter is added to qps_open, which can be used to only retrieve
the metadata details of the store, without mapping all the QPS pages.
Opening a QPS in such a way is very dangerous, as the object won't be
properly initialized, and almost all QPS API calls will be invalid.
The QPS object must be immediately closed afterwards.

It would be much better to add a new API that returns the metadata but
does not return the QPS object. This requires some changes by the
callers however, to not depend on the QPS object anymore.

Change-Id: I56a7d4c05ce646a2eaa8af90ea941ec8e03007f2