iop-yaml: properly handle & test tricky strings
Due to YAML handling of non-quoted strings, it is possible to have
strings that will be packed incorrectly as they will map on other types.
For example, "~", if left unquoted, will be parsed as the null value.

For this reason, the best behavior possible is to always quote strings
if, when parsed, they are going to be parsed as another type. This can
however prove very costly to test, as it depends on the behavior of
strtol, strtol etc.

Instead, the behavior chosen is the following:
* For ~ or null, quote it. This is necessary because a null value can
  have a very different behavior from other types.
* For true/false, do not quote it. This means the string will be parsed
  as a boolean. However, when interpreting the AST into an IOP object,
  if the field is a string, we can get back the string from the input
  stream.
* Same thing is done for double and uint. It is not necessary for "int"
  for the moment, as the presence of a "-" symbol will automatically
  quote the string.

This is a bit of a weird behavior but it is fine because the AST is only
used to be interpreted as an IOP object afterwards, and we can always
get back the exact string from the input stream.

Change-Id: I9930b3fe35863fb4bd202875fd4c8e2a32b5f308