iop-yaml: improve error msgs on struct unpacking
The call to the YAML unpack function always added a "cannot unpack as "
to the error message. However, if the error was not inside a field of
the struct, the yaml_data_to_typed_struct also added such a message,
leading to an unneeded duplication.

This is simplified by always adding this message in
yaml_data_to_typed_struct, and not adding anything at the top-level.
Idem for union.

Change-Id: I5eda26f23ef77600aa36253a840e7d2444b93ba4