iop-rpc: factorize retrieval of details from ic data
The code is very complicated due to a high combination of possibilities,
and functions used for differents purposes.

What we have is basically:
* for normal cases, we want to get the hdr & the value from the msg,
  to be able to do the pre hook and call the RPC callback.
* for static proxy cases, we only use the hdr if there is a pre hook
* for dynamic proxy cases, we need the hdr and the header len for
  repacking purposes.

Each case was doing its own call to t_get_hdr_value_of_query, which was
further complicated by the pre hook case.

Now, we do this call before anything else. This factorize the retrieval
of the hdr, but it makes which variables are set more complex in the
main function (hlen, hdr & value might be set or not depending on the
type of the callback).

Change-Id: I38a9c9fd63149e059f2c1b6152694720cadfc598