iop: rework function 'iop_value_set_here()'
Seen while reviewing iop-yaml code: the function name was really unclear
and the function was some kind of an hybrid supposing that the field was
optional in some cases and supporting the references in the other hands.

This function is renamed as 'iop_field_set_present()' (as a function
'iop_field_set_absent()' already exists) and a new function
'iop_field_alloc()' is added specifically for optional structs/unions or
references.

This patch comes with a important amount of assertions that may help
detecting errors while modifying the code thus increasing code
maintainability.

Change-Id: I7d18c42ef488020ffb00fd72a7d9a84e8622021c