Remove dangerous if_assign macro (and unused while_assign one).
This macro was pretty convenient, but was actually really dangerous
because of the hidden for it used.

Let's take this code:

 for (...) {
     if_assign (x, y) {
        ...
     } else {
         continue;
     }
     ...
 }

Because of the hidden for in if_assign, the continue actually applied to
it, and not the first one, and thus had no effect.

After a long internal debate at Intersec, it was decided to just get rid
of this dangerous helper to avoid stupid bugs like that.

Change-Id: I3ec8c8a2c376c2e5b10a37d2f08de745a58900df
Refs: #69303