sctp: Add SCTP minimal API header file.
The libsctp-dev provides a header file (netinet/sctp.h) which contains the
structures, the constants and the function's prototypes of the libsctp
user space API.
However, this library is not used as some of its functions were
re-implemented in net-sctp.c. So the header file was used to only get the
structures and constants definition.

But it seems that the libsctp-dev header file contains some incoherency.
For example the enum sctp_spinfo_state differs from the kernel one.
Then some bugs may occur when using the libsctp-dev header file.

A minimal header file is now created and contains only the structures, the
constants and the function's prototypes which are not defined in the
kernel sctp header.
The linux-uapi-sctp.h corresponds to the kernel file
include/net/sctp/user.h from kernel version 2.6.32.

The linux kernel uapi sctp header file is then used to access to the
remaining structures and constants definition.
However for kernel version < 3.10, this uapi header doesn't exist. So a
new header file (similar to linux uapi sctp header) is created.
This new header (linux-uapi-sctp.h) is included only if
/usr/includelinux/sctp.h doesn't exist.

Be carrefull, the linux-uapi-sctp.h corresponds to the kernel header file
(include/net/sctp/user.h). For this version SCPT_ACTIVE has the value 1.
This value has changed in later version.

PS: The preprocessor HAVE_NETINET_SCTP_H is not usefull now as we don't
need the installation of the libsctp-dev package. This preprocessor is
then removed.

Conflicts:
	Build/configure.inc.sh

(cherry picked from commit 442b97f7fdee9757687304f34a1d74ec326fc4be)

Refs: #64114
Refs: #72386
Change-Id: Ia64c22a919a210a938e90d7c53575450d2f1c596