yaml: replace the key->data map in yaml_obj_t with a vector
Instead of using a map for the fields of a yaml object, use a vector.
This is needed to keep an ordering of the fields, which will be required
in order to implement the AST -> YAML conversion.
This makes the AST -> IOP part slightly worse, as we now do a linear
search on the AST to find if it mentions a field. For the moment, this
is considered ok as the focus is not on performances for the yaml
serializer.

Change-Id: I36e86f5fdefe409f9e29b77580c193346cdfe35f