iop-openapi: handle union types
IOP union types are discrimated with a field name. To match that on
OpenAPI, there are two possibilities:

* use a standard "object" description (with every field described in the
  properties field), but make them all optional (no required field) and
  add minProperties = maxProperties = 1.
* use oneOf of different "dummy objects", each containing a single field
  of the union.

The first solution would probably generate very confusing indications to
the end-user, while the second one describes in a much clearer way how
to declare the union field, so the second solution is picked.

Refs: #69615
Change-Id: I05802f0b36cec9c48b13c0f9cf95e37381f4a52c