git: Ignore gcc temporary files.
When compiling, gcc creates temporary files (with .c.tmp extension) in the
source directory. Because we were not ignoring them, the version.sh
script, if called at the wrong time, could generate a file with a dirty
revision while it was not justified.

Fix that by adding gcc temporary files in .gitignore.

I've also opened the following waf ticket about this:

 https://gitlab.com/ita1024/waf/issues/2267

Change-Id: I34cd769ea030fb9910f9b258a1a802e5ad42cf99
Closes: #69154