iop-openapi: add paths from IOP RPCs
Starting from an IOP "module", we add all RPCs in their own path.

The path is constructed as "/{product}/{X}/{rpc}", where X is the name
of the interface in the IOP module.

There are a lot of things to fix, as can be seen in the test results,
but this is a good start.

Refs: #69615
Change-Id: If6888768906f4420a918a75fe63149957f35bc1d