iop-openapi: forbid addition of multiple IOP modules
Due to how IOP modules work, multiple IOP modules can expose the same
interfaces, but the REST routes will only depend on a single module (as
it uses the interfaces aliases of the module). So an OpenAPI description
should only be created from a single module.

To fix the iop-openapi API, the IOP module is moved as an argument of
t_new_iop_openapi. Only a single one can thus be used.

To keep the tests simples, the argument is kept nullable, so that we can
keep tests that do not depend on any IOP module.
This also forces us to change the "clear schemas" hack for one test, as
the schemas are now generated in the "to_yaml" function.
This is fixed by fixing up the generated yaml, instead of being
intrusive in the iop-openapi API.

Refs: #69615
Change-Id: Ic78c2142d15ec4023a8bc2954bdb2f8f0cfae57e