iop-openapi: fix tricky wrong pointer use
In the union -> schema function, we first build a new schema object and
keep a pointer on it. We then add new schemas for every field of the
union. These additions can trigger a resize/move of the vector,
rendering our first pointer invalid.
To fix this, only add the schema in the schemas vector at the end of the
function.

Also, remove a t_scope that is not required, to avoid resizing t_scope
allocated objects while a new t_scope was declared.

Refs: #69615
Change-Id: I084d560002fae982f8a0f4f8394a668053cef0c7