zchk-snmp: Fix smilint warnings.
Previous commit introduced a regression in the test
iop_snmp_mib.test_entire_mib: smilint was printing the following warnings:

 test-data/snmp/mibs/Refs:-INTERSEC-MIB.txt: No such file or directory
 [1] {module-not-found} failed to locate MIB module `INTERSEC-MIB'
 [1] {object-identifier-unknown} unknown object identifier label `test'

This is because the relative path to REF-INTERSEC-MIB.txt given as smilint
argument was not good anymore (it seems that the programs are executed
from the lib-common directory wherever the zchk is when running
`waf check`).

Fix that by using an absolute path.

Sadly, smilint was still returning 0, so the error was not catch. I'll fix
it in the next commit.

Change-Id: I8216486d9d5987d53e1b1cb52066140934db1da3