iop-yaml: remove generic mem pool, only use t_scope
The code of the YAML unpacking into AST and then into IOP uses a generic
mem_pool (which of course must be frame-based).
However, only the t_pool can be used with the public API, as the code
was never tested with any other mempool.
As the support of other mempools is not tested, not supported and
probably buggy, and as the use-cases for other mem pools is not clear at
all, the code is modified to use the t_pool explicitly, which helps
making it a bit simpler and safer.

If in the future, use-cases with other mem pools are found, the code can
be modified to become generic, and then properly tested with other
mem-pools. This will also help simplifying the future feature of
metadata saving.

Change-Id: I627e72280a5f730fe6809396cb54264e1b7191cb