yaml: parse flow syntax for sequences
YAML allow writing sequences in "flow" format, that is with '[' and ']'
delimiters, using ',' as a separator.

In a flow context, however, parsing is more limited: there is not tag,
no classic obj or seq syntax, the only data allowed is:
* scalars, but ']' and ',' are now separators for strings
* implicit objects: a 'key: val' data is taken as an object with a
  single mapping.
* flow objects (not handled yet)
* flow sequences
In a flow context, there is no notion of indentation as well.

Change-Id: I8d060134c6ddb3a8fc554bd9c9d4e7d138b760c7