yaml: fix parsing of comments containing '#'
If a comment contains the # character, for example for a ticket number,
the comment was badly saved in the presentation data, leading to
only the part after the # character being saved. So when repacking,
the YAML went from:
 a # comment # foo
to:
 a # foo

This is fixed by only setting comment_ps if we were not in a comment
context already.

Change-Id: I0170792e9964ee667cbb7e8aa137dd8aeca35ce8