yaml: save empty lines in presentation data
It is common in YAML documents to add newlines to make the document
breathe a bit more.
It is usually done on the keys of the root data:

 a:
   z: 1
   ...

 b: ...

But can be done anywhere. This information is saved in presentation data
and used when repacking. Empty lines are added before the prefix
comments.

Change-Id: Ib56db06402cfb00e2d4b8374e5572e8c2bd28016