yaml: add path to file being parsed in t_yaml_parse
This parameter will be used for two things:
* use the filename in the error message for clearer error reporting.
* use the filepath to compute paths to files being included, relative to
  the file being currently parsed.

We have to pass the pstream_t of the file + the filepath to t_yaml_parse
instead of having a t_yaml_parse_file. The reason for path is that
the yaml_data_t object is keeping references to the parsed stream, so
the parsed stream must outlive the use of the yaml_data_t object.

Change-Id: I8a0d979836a5fdedeee8ed4fba6a60153475b154