bit-buf: fix assert in bb_init_full with an empty sb
sb_init(&sb) + bb_init_sb(&bb, &sb) triggered an assert in sb_grow because
__sb_slop is not aligned on 8 bytes:
 mcms-cbc: platform/lib-common/bit-buf.h:89: bb_init_full: Assertion
 `((intptr_t)buf) % alignment == 0' failed.

To fix this problem, add a special handling for __sb_slop.
Also fix a memory corruption when sb->size is not a multiple of 8.

This assert was triggered by mcms-cbc with sabp protocol.

This patch could be backported in 2013.3.

(cherry picked from commit eefb2ebf5f97721ddd50a575c9d08df7c170456c)
backport to 2016.4 because it's the first branch where it does not
conflict.

Change-Id: I2868915cecb9b9ce7797270d7d520281ed4de644
Refs: #70166