yaml: consider scalars for presentation
Add presentation data to scalars. Having only presentation data for keys
or sequences indexes was buggy from the beginning, as a log of syntaxes
were either ambiguous or would just associate comments to an unexpected
data. In addition, there was a hack to handle presentation data for a
root scalar, but not others, complexifying the code.

Now, scalars can also have presentation data, and are referenced with a
'!' in the presentation path.

Change-Id: I1530924002bb246ffeaeec6ff46ad95cd872f9d2