yaml: add yaml_parse_t object to simplify parsing APIs
Add a top-level object to use when parsing. This will allow having a
file parsing API, as well as adding helpers to set parameters without
having to specify them all in the "parse" call.

We lost the filename in error messages, but it will be added back in
the next commit.

Change-Id: I9f41083b930906cd4c9318c18cd7d789bf3d42f9