yaml: add proper error buffer to packing API
Add an error buffer to the yaml_pack API, and to the write callback,
so that if an error is returned, we ensure an error buffer is filled.
For the moment, only the write callback can fail, but in the future,
the packing itself will be able to fail (notably, when trying to
recreate included subfiles).

Change-Id: If56c0a22a0081a99f66a24725f958d54c203cefa