waf: Ensure Iop2c tasks are executed before Blk2c/Blkk2cc/ClangCheck ones.
.blk files can include .h files that are produced by Iop2c tasks, and it
is not detected by waf because the Blk2c tasks ae not using the C waf
preprocessor. Because of this, there is a race that can make the builds
fail.

In order to fix that, I first thought to use the waf C preprocessor as the
scan method of the Blk2c class, but I finally decided to not do it
because:
* the scan would be done twice per .blk files when compiling with gcc
  (once for the .blk itself, once for the .blk.c).
* the Blk2c task would be re-executed if the content of an included .h
  changes, which is useless.

Instead, just use a 'before' constraint to guarantee the tasks order, so
that all Iop2c tasks are executed before all Blk2c tasks. This is already
done for Fc2c task for example.

Change-Id: Ie6a331bf15eb9fce17fcbde2c21c9f840c5b7e6c