iop-rpc-channel: fix the use of erroneous IC hook context pointer.
After the pre_hook, the IOP RPC channel layer checks the global
variable _G.ic_hook_ctx to know if there is an allocated context (i.e
the pre_hook has succeeded). If so, the callback and post_hook
associated to the RPC are called.

This _G.ic_hook_ctx variable is only set in ic_hook_ctx_new, and unset
in ic_hook_ctx_delete:
- ic_hook_ctx_new is supposed to be called in the pre_hook.
- ic_hook_ctx_delete, is called in the IC reply.

So, If the pre_hook fails and ic_hook_ctx_new/ic_hook_ctx_delete are
not called inside the pre_hook, which is a legit behavior, the variable
_G.ic_hook_ctx is not modified.

This will lead to a crash because the pre_hook error is not detected in
ic_query_do_pre_hook. For example:
- We first have a query that is correctly authenticated, and
_G.ic_hook_ctx is set. This RPC triggers a behavior with asynchronous
reply, so we do not unset _G.ic_hook_ctx immediately.
- Before the async reply, an unauthenticated request is received, and
execute the pre_hook, but _G.ic_hook_ctx is not modified since
ic_hook_ctx_new is not called.
- After the pre_hook, _G.ic_hook_ctx still contains the context of the
first requests. The function ic_query_do_pre_hook returns 0 instead of
-1 and we execute the callback of the associated RPC which should not be
the case.
- This leads to a crash because we release too many times the HTTP
query (or some other odd behavior).

This is fixed here by saving the context before the pre_hook, and
setting _G.ic_hook_ctx to NULL.

Change-Id: Ice8653b008f429b7bb5e4fb2c9befe04a856bd50
Refs: #70264