yaml: start handling includes in presentation
Save the information that a yaml data was included from another file
in the presentation node of this data. When repacking, create a new
subfile inside the output directory, pack the data in it with its
presentation details, and then write the "!include <path>" data.

This is still a bit rough, as evidenced by the added FIXMEs and the lack
of checks in some parts of the code. This will be improved in the
following commits.

Also, the z_t_yaml_test_parse_success was returning a yaml data but
cleaning the parse env, which can lead to some invalid reads if the data
contained included files and is then used. To fix this, the parse env
can be returned as well, so the caller can decide when to delete it.

Change-Id: I613c8e6d62c8aabfd0e4e634cf7eafe300d3ae4d