yaml: fix presentation with flow mode
The handling of comments with flow mode was quite buggy. Flow mode is
handled in a specific way for parsing, and we do not save any
presentation data when parsing it. This however was not properly handled
when packing it, leading to either assertion failed on multiple inline
comments, or buggy insertion of comments.

To fix this:
* parse the whole flow data as a "scalar", so all presentation data will
  be stored in the "!" key.
* when packing, get the data from the "!" key to add inline/prefix
  comments.
* make sure the pack_scalar, which can be called in flow and std
  context,  does not try to get presentation data in flow mode.

The code is starting to get more complex as it goes on, but unless the
presentation rules that are saved are simplified, there is no real way
around it, as the way yaml is packed makes it non-trivial to save &
restore contextual data.

Change-Id: I77455797c738e8b57f5f19a912ed1e8c72f0bab0