yaml: include presentation data in yaml_data_t object
Instead of having a "current_path" when parsing, and trying to generate
the presentation data outside of the generated data objects, split this
in two steps:

* parse objects, and associate presentation data to:
  * yaml_data_t object
  * seq elements (ie to the '-' character)
  * key of objects (ie to the '<key>:' string)
* then, generate a presentation_t object matching paths to presentation
  nodes.

While this forces us to go through the generated AST again, it
simplifies the code and removes some of the ugly code that were added in
the last commits. We now only have to handle two states:
* the last parsed element that can have a presentation data
* presentation data saved for the next element.

There is also no complex path_len/push_pop path inside the parsing,
making the code easier to read.

This also removes some hacks, for example with flow mode, or for
presentation that could be associated with complex data. This makes the
repacking more stable, also the association with paths can now be a bit
more surprising.
For example, for this example:

 # A
 a: 1
 # B
 b: 2

"A" will be associated with the object itself instead of the first key
(ie with path '!').
However, "B" will be associated with the second key (ie with path ".b").

Change-Id: I0530737f07497d4b33c3eea59bfc949e88f8954d