yaml: detect and handle conflicts when packing shared subfiles
It is possible to include the same files multiple time in a YAML
document. We want to at least memorize which files were written, so
that we avoid rewriting the same file again and again.

However, there is also another issue, in that the data may have been
modified, which means that we should pack those data in different files,
as they can no longer be shared in the same file.

This is handled by keeping a context of fullpath to the subfile, and
associated YAML data. When packing a subfile, we detect if it was
already written, and compare the YAML data with the one already packed.
Then, we add a ~%d suffix until we find an available file.

There is still some improvements required on path handling, as we really
need to make sure all paths are canonical, to prevent accidental
rewriting.

Change-Id: I52def4ca7c1ad2fe0adad32eaa2756484e6086aa