yaml: simplify path when repacking includes
In order to properly detect identical included files but with different
paths, the presentation path is simplified. This ensures that the path
to a subfile is unique, because:
* it must be inside the output directory of the including file
* the packing does not create symbolic link, so no need to call
  "realpath".

Technically, since we do not forbid packing in an existing directory, it
is possible for the user to pack in a directory where symbolic links are
present, which would mess up the subfiles handling logic.
This is left as a known buggy situation, but handling this properly
either means rejecting packing in existing directories, and adding lots
of complex code for it...

Change-Id: I7a39c8701a6d093b4a6393d1dc3f651ac2f39e1f