yaml: remove allocation of document presentation on retrieval
Now that the document presentation struct is public, there is no longer
a need to have an extra indirection to keep the struct private. Now
callers can declare a variable of the struct type directly, and have it
filled by the get_presentation call.

However, the double pointer is still useful for IOP unpacking, as the
presentation needs to be dup'ed before it is returned. The const is kept
for now as we don't expect the user to modify the presentation by hand,
but it could be removed in the future.

Change-Id: I2d094da61e07e365ed669d102f733f27c5110090