iop-yaml: Add helpers to convert IOP JSON to YAML presentation
In order to be able to convert existing JSON documents into YAML, we
want to preserve the includes. To do so, the JSON subfiles must be
converted into a YAML document presentation.

This is pretty straightforward, expect for the fact that the data has to
be converted from a flat representation to a layered one. As the data
from the JSON unpacking is properly ordered (ie, subfiles included from
another subfiles always follow the including subfile in the array), the
conversion is not complex.

Change-Id: I1482966b4a3d777bd91debab30a0afca3e0b9c71