yaml: add start of override implementation
On specific operations, such as includes, we want to be able to
overwrite the included AST with our own values.
This is done by specifying an override object just after the include,
which is then merged with the included AST.

The main problematic is how to do the merge: adding to the existing AST,
or replacing it.

For example, official YAML supports anchors, with overrides, however
replacing values. This makes adding to the AST extremely painful:

a: &anchor
  b:
    c: 1
    d: 2

z:
  <<: *anchor
  b:
    c: 3

will return:

z:
  b:
    c: 3

and the d node is lost.

Here, for the moment, we only handle adding values. There are two open
questions to support the rest:
 * which syntax to specify replace
 * how to refer to specific sequences elements, for example to override
   a specific element.

And of course, the presentation handling of overrides will be far from
trivial.

Change-Id: I4d6953404935b23a53c83a1f6f8c543078e4db1c