yaml: handle added nodes in override presentation
Some data types are added by override to the overridden AST, instead of
simply replacing the original value. This is the case for sequence
elements, as well as unmatched keys.
To handle this in presentation, those paths are stored in the
presentation override object, but without an associated original data.
This indicates there was no original data.
Then, when repacking, those nodes are ignored when packing the AST, and
pushed in the override object.

Change-Id: I262701415bcd403795b1aadc4a5baed4b9954574