yaml: handle modifications of AST modified with overrides
Fix and test how changes in an AST that was created through overrides is
handled.

We have two situations:
First, a data that was set by an override is modified. In that case, we
want the new value to still be in the override (ie, the override is
modified, not the included data).

For example:

in.yml:
  a: 1

out.yml:
  !include in.yml
  a: 2

If the AST is modified, and the repacked data is "a: 3", we want:

in.yml:
  a: 1

out.yml:
  !include in.yml
  a: 3

This was already properly considered in the design of the repacking of
overrides, and works well.

The second situation is if an overrided node is removed from the AST.
This will lead to an override not being found when repacking the inner
AST, and must be detected. To do so, a new "found" boolean is required
in our repacking override data.

Change-Id: I51dd6d6eb039e4aa92a0df4135a17e835963b06e