yaml: compare shared subfiles with contents' checksum
When a subfile is included in multiple places, we need to ensure the
repacked data is the same on all those places. Otherwise, we must
rename the subfile to prevent altering the packed AST.

This detection was previously done by comparing the two ASTs being
packed in the same file.
This is no longer correct, for multiple reasons:
* this was written when the presentation data was private, and could not
  be modified by the user. This is no longer the case, and presentation
  data impacts the packed document, so it should be considered to
  compare the two packed data, simply comparing the AST is no longer
  valid.
* The override presentation is implemented in a way that requires going
  through the overridden data to properly:
   * retrieve the original data and pack it
   * save back the actual AST into the override data node.
  This means packing a subfile has actual side effects on the current
  document, and must be considered to compare the shared subfiles (ie,
  two different ASTs could be packed in the same shared subfiles, if
  the differences are offload to an including override).

Those differences means it is very hard to properly test equality of
shared subfiles in advance. To fix this, we now actually do the packing,
then compare a checksum of the contents that would be written in the
subfile.

This isn't exactly ideal as we use a temporary buffer before writing it
all in a file, but this is much safer than the previous method.

Change-Id: Ida8b7b239d0829aa1a2a867bdfbd1860fd940ac2