yaml: make yaml_pack_sb failable
yaml_pack_sb was implemented as being unable to fail, but it isn't true.
It is actually possible to define an output directory for the packing
environment, and then pack in a sb. That will recreate all subfiles
(which means it can fail), but will pack the root file in a sb buffer.

This is actually a use-case we will want to use internally, as well as
useful for tests. This is now properly handled:
* yaml_pack_sb can fail, and the documentation is updated to explain
  when it can fail.
* the initialization of env->subfiles is made lazy to be able to work
  with yaml_pack_sb.
* the new use-case is used in tests to ensure it works.

Change-Id: I4ec14dd4d12cbbfc4fcaf5645079e188e28b4f3e