core/types: improve semantics of REFCNT generated functions
Rename 'dup' as 'retain'. The naming was abusive because the function
didn't actually duplicate anything.

Add a new 'release' function that deletes the object if the ref count
reaches zero but does not erase the pointer value if there are still
references on the object. This is quite useful when we want to protect an
object against deletion by using 'retain/release' around some user code.

This renaming makes the DO_REFCNT functions consistent with the similar
core object methods.

Change-Id: I1a39eabece80e7bd9f2b4c5cac29555975301b8e