yamlfmt: allow repacking with subfiles
Add an option to output into files, recreating subfiles. In addition,
it handles conversion from JSON with subfiles to YAML with subfiles.

The ugly thing here is that we would want to only specify the output
directory, but we actually need also the name of the output file, as
choosing one ourselves could lead to conflicts with subfiles.

Change-Id: I874f65fd350e7e4dd15a09d8841fd9488f34ca72