yaml: improve handling of NO_SUBFILES case
Remove the number of "if (no_subfiles)" specific code, by making
the whole path_to_checksum + check_subfile code executed in the
no_subfiles case as well. This does not change anything for this
particular case as the checksum will always be the same, but it makes
the code more readable.
Also, make outdirpath always set at least to "". The "pack into files"
test is update to reflect this.

Change-Id: Ief3b36f885481d8fb0e9c90fb51fceff3a0fa9e5