iop: set parameter 'sdesc' in 'iop_skip_absent_field_desc()' as optional
In some use cases of 'iop_skip_absent_field_desc', it's more practical to
avoid having to keep the struct description along with the field
description.

The good news is: we don't really need it. Set it as "nullable" so the
caller can skip it if he don't want it.

Change-Id: I37c45d87b0bddaca3360c53b60019d1fcc1d1478