yaml: save quoted state in presentation data
Save information about whether a string was quoted or not in
presentation data.
This will improve the stability of using quoted strings that are not
required to be quoted. This will also help in cases where special values
were quoted (for example true or false). Finally, this will be useful
to handle escaped variables in quoted strings in the following commits.

Change-Id: Ib7f215711c6e6f88252a3836be78f3750879f0c6