parsing-helpers: only expose parse_backslash helper
The function parse_quoted_string was added in an effort to factorize
code between yaml & json parsing, and avoid duplicating handling of
escape sequences.
However, it gets tiring to have to maintain compat with the JSON code,
and it won't be possible to do so if we want to handle variable escaping
in YAML.
So only factorize the parse_backslash helper, and put the json code
back as it was before the factorization.

This actually improves the YAML code as well, as shown by the TODO
removed in a test.

Change-Id: I80705ed5c2d5d1b01554af09dc5bccade72a85b1