yaml: handle escaping for variables
Escaping variables turned out quite harder than expected. There
are several issues that need to be handled for it to work.

Escaping means that when parsing:

a: "a b \$c d"

we properly have the string "a $b c" in the end, without errors about
unbound variables.

The main issue is when using strings that have both variables and
escaped dollars:

a: "a $b \$c $d"

We have several choices in this case:
* keep the string raw, so that when replacing variables, we detect which
  symbols are escaped. This is actually quite hard to do, it requires
  not evaluating the string at all (to avoid evaluating "\\$" as "\$"
  for example), and keeping the information about whether it has been
  cleaned or not. In addition, it isn't as simple to properly repack the
  string if we only substituted some variables, but not all. Plus, it
  requires sanitizing the AST before returning it to the user.
* evaluate the escaping when parsing, and keep the information about
  which '$' are variables and which are not. This means keeping a bitmap
  of this information with us all the time, which is a bit cumbersome:
  * it must be created on evaluation, and kept available
  * it must be updated on every partial substitution
  * it must be used everywhere we use the template string
* divide the string in chunks: ["a ", var:b, "$c ", var:d]. This is
  probably the cleanest, but it is not cheap, and falls on the same
  issues as the first one, where we keep an intermediate representation
  that needs canonization before return to the user.

So the second solution has been kept. To keep the classic case fast,
this bitmap is only used if a string contains both variables and escaped
dollar characters. Only parsing is handled, the packing is done in the
next commit.

There is still a big issue, which will require its own fix: if the value
of a variable contains '$' characters, the bitmap must be fixed, to
prevent it from not matching the string. This is quite a cornercase, but
it could lead to UB memory read.

Change-Id: I9d7e7841fb3bcfa5fa30f626cf64b06bba08cc38